+++ vello/src/lib.rs
@@ -147,13 +147,12 @@ pub mod low_level {
 pub use peniko;
 /// 2D geometry, with a focus on curves.
 pub use peniko::kurbo;
-pub use skrifa;
 
 #[cfg(feature = "wgpu")]
 pub use wgpu;
 
 pub use scene::{DrawGlyphs, Scene};
-pub use vello_encoding::Glyph;
+pub use vello_encoding::{Glyph, NormalizedCoord};
 
 use low_level::ShaderId;
 #[cfg(feature = "wgpu")]